backend/requirement.txt
annotated-types==0.7.0
anyio==4.13.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
distro==1.9.0
filetype==1.2.0
google-auth==2.49.1
google-genai==1.69.0
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.23
langchain-google-genai==4.2.1
langsmith==0.7.22
numpy==2.4.4
orjson==3.11.7
packaging==26.0
pandas==3.0.1
protobuf==7.34.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pymilvus==2.6.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyyaml==6.0.3
requests==2.33.0
requests-toolbelt==1.0.0
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
langchain-text-splitters
langchain-google-vertexai